requirements.txt
streamlit
pymupdf
strands-sdk
matplotlib
pillow
